+++ config/custom_func.go
@@ -5,6 +5,7 @@ type ArrayCustomFunc func(a []any, args ...any) []any
 type IntCustomFunc func(i int, args ...any) int
 type FloatCustomFunc func(f float64, args ...any) float64
 type BoolCustomFunc func(b bool, args ...any) bool
+type ObjectCustomFunc func(o map[string]any, args ...any) any
 
 type Func struct {
 	Str   map[string]StrCustomFunc
@@ -12,6 +13,7 @@ type Func struct {
 	Int   map[string]IntCustomFunc
 	Float map[string]FloatCustomFunc
 	Bool  map[string]BoolCustomFunc
+	Obj   map[string]ObjectCustomFunc
 }
 
 func NewFunc() *Func {
@@ -21,5 +23,6 @@ func NewFunc() *Func {
 		Int:   map[string]IntCustomFunc{},
 		Float: map[string]FloatCustomFunc{},
 		Bool:  map[string]BoolCustomFunc{},
+		Obj:   map[string]ObjectCustomFunc{},
 	}
 }