+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - ⚠️ BREAKING CHANGE! Rename functions inside a parser. It should not be a breaking change unless you use Parser directly in some rare cases.
 - ⚠️ BREAKING CHANGE! Minimal Go version support is version `1.25`.
 - 🧑‍💻 Improve error handling when trying to use `@use`, `@insert`, `@reserve` or `@component` directives in simple `EvaluateString` or `EvaluateFile` function calls. These directives are only allowed inside template files with `textwire.NewTemplate`.
+- ✨ Now you can add custom functions to objects as well with `RegisterObjFunc`.
 
 ## v2.7.1 (2026-01-23)
 - 🐛 Fixed incorrect file path in error messages when error happens inside of components